+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - MySQL/PostgreSQL: Increase maximum length limit for addressbook name (Fixes #382)
 - Fix: log messages could go to the wrong logger (carddav\_http.log) for a small part of the init code
+- Support setting roundcube's collected senders/recipients to addressbooks from preset (Fixes #383)
 
 ## Version 4.3.0 (to 4.2.2)